refactor(frontend): clarify accessible color utilities
